refactor: higher-level client definition
